更新管理员页面
