+++ src/views/admin/TaskRet.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import { ref, onMounted } from 'vue'
-import { type taskItem, getTaskList, adminAuth } from '@/api/admin'
+import { ref, onMounted, watch } from 'vue'
+import { type tasrRetItem, getTaskRet, adminAuth } from '@/api/admin'
 import { useRouter } from 'vue-router'
 import { useUUIDStore } from '@/stores/userInfo'
 
@@ -8,23 +8,39 @@ import { useUUIDStore } from '@/stores/userInfo'
 const store = useUUIDStore()
 //route
 const router = useRouter()
-
+const uuid = store.getUUID
 //页面加载样式
 const dataListLoading = ref(true)
+const dataListShow = ref(false)
 
 //身份验证数据加载
-const taskList = ref<taskItem[]>([])
+const taskList = ref<tasrRetItem[]>([])
+
+//分页数据偏移量
+const currentPage = ref(1)
+
+//监听分页
+watch(currentPage, () => {
+  dataListShow.value = false
+  dataListLoading.value = true
+  //更新数据
+  getTaskRet(currentPage.value - 1, uuid).then((rep) => {
+    taskList.value = rep
+    dataListLoading.value = false
+    dataListShow.value = true
+  })
+})
 
 //挂载事件
 onMounted(() => {
-  const uuid = store.getUUID
   if (uuid !== '') {
     adminAuth(uuid).then((rep) => {
       if (rep === true) {
         //更新数据
-        getTaskList(uuid).then((rep) => {
+        getTaskRet(currentPage.value - 1, uuid).then((rep) => {
           taskList.value = rep
           dataListLoading.value = false
+          dataListShow.value = true
         })
       } else {
         router.push('/error')
@@ -37,53 +53,33 @@ onMounted(() => {
 
 const statusList = ref([
   {
-    text: 'close',
-    type: 'warning'
-  },
-  {
-    text: 'ready',
-    type: 'primary'
-  },
-  {
-    text: 'ready',
-    type: 'primary'
-  },
-  {
-    text: 'ready',
-    type: 'primary'
-  },
-  {
-    text: 'lose',
+    text: 'fail',
     type: 'danger'
+  },
+  {
+    text: 'success',
+    type: 'success'
   }
 ])
 
 const clickBack = () => history.back()
-
-function goTaskRetPage() {
-  router.push('/mario/taskRets')
-}
 </script>
 
 <template>
   <div class="body">
-    <van-nav-bar left-text="返回" left-arrow @click-left="clickBack">
-      <template #right>
-        <div class="goTaskArrowWrap" @click="goTaskRetPage">
-          <van-icon name="arrow" size="18" />
-        </div>
-      </template>
-    </van-nav-bar>
+    <van-nav-bar left-text="返回" left-arrow @click-left="clickBack" />
+    <van-pagination v-model="currentPage" :total-items="24" :items-per-page="5" />
+
     <!-- 加载 -->
     <div class="loadingBox" v-show="dataListLoading">
       <van-loading size="30" color="#1989fa" />
     </div>
-    <div class="taskListWrap">
+    <div class="taskListWrap" v-show="dataListShow">
       <van-cell-group>
         <van-cell
           v-for="(task, index) in taskList"
           :title="task.username"
-          :label="task.add_time"
+          :label="task.time"
           :key="index"
           center
         >
@@ -111,10 +107,5 @@ function goTaskRetPage() {
   display: flex;
   justify-content: center;
   align-items: center;
-}
-.goTaskArrowWrap {
-  display: flex;
-  align-items: center;
-  justify-content: center;
 }
 </style>
